test(point-cloud-layer): colocate antialiasing render coverage
